proof of concept fix of #17194.
Makes both ActionController's redirect_to and ActionDispatch's redirect
respond with a body format that honors the requests Content-Type of
the issuing request.

The code sharing here is logical, but probably in the wrong place.

Missing tests; awaiting feedback before diving beyond this proof of
concept.

A 4.2.3-compatible monkey patch is available here:
https://gist.github.com/christhekeele/f56e6512d92ed3af0427